Move RandomState and DefaultHasher into std::hash, but don't export for now
